chore: move the program dashboard APIs
Moves the Program Dashboard APIs out of the deprecated remnants of the
legacy learner dashboard, into the Programs djangoapp.

Keeps the old legacy routes for this API, left over from the deprecated
remnants of the legacy learner dashboard, alongside future-proofed
routes which will work when the deprecated, legacy Program Dashboard is
replaced with functionality in the Learner Dashboard MFE.

FIXES: APER-3949
