+++ lms/djangoapps/learner_home/rest_api/urls.py
@@ -1,0 +1,11 @@
+"""
+Programs API URLs.
+"""
+
+from django.urls import include, path
+
+from openedx.core.djangoapps.programs.rest_api.v1 import urls as v1_programs_rest_api_urls
+
+urlpatterns = [
+    path("v1/", include((v1_programs_rest_api_urls, "v1"))),
+]
