+++ lms/djangoapps/learner_home/urls.py
@@ -6,6 +6,8 @@ from django.urls import path
 from django.urls import include, re_path
 
 from lms.djangoapps.learner_home import views
+from .rest_api import urls as rest_api_urls
+
 
 app_name = "learner_home"
 
@@ -13,4 +15,5 @@ app_name = "learner_home"
 urlpatterns = [
     re_path(r"^init/?", views.InitializeView.as_view(), name="initialize"),
     path("mock/", include("lms.djangoapps.learner_home.mock.urls")),
+    path("", include(rest_api_urls)),
 ]
