+++ openedx/core/djangoapps/programs/rest_api/v1/tests/test_views.py
@@ -8,10 +8,6 @@ from uuid import uuid4
 from django.core.cache import cache
 from django.urls import reverse_lazy
 from enterprise.models import EnterpriseCourseEnrollment
-from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
-from xmodule.modulestore.tests.factories import (
-    CourseFactory as ModuleStoreCourseFactory,
-)
 
 from common.djangoapps.student.tests.factories import (
     CourseEnrollmentFactory,
@@ -40,17 +36,22 @@ from openedx.features.enterprise_support.tests.factories import (
     EnterpriseCustomerFactory,
     EnterpriseCustomerUserFactory,
 )
+from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
+from xmodule.modulestore.tests.factories import (
+    CourseFactory as ModuleStoreCourseFactory,
+)
 
-PROGRAMS_UTILS_MODULE = 'openedx.core.djangoapps.programs.utils'
+PROGRAMS_UTILS_MODULE = "openedx.core.djangoapps.programs.utils"
 
 
 @skip_unless_lms
-@mock.patch(PROGRAMS_UTILS_MODULE + '.get_pathways')
-@mock.patch(PROGRAMS_UTILS_MODULE + '.get_programs')
+@mock.patch(PROGRAMS_UTILS_MODULE + ".get_pathways")
+@mock.patch(PROGRAMS_UTILS_MODULE + ".get_programs")
 class TestProgramProgressDetailView(ProgramsApiConfigMixin, SharedModuleStoreTestCase):
     """Unit tests for the program progress detail page."""
+
     program_uuid = str(uuid4())
-    url = reverse_lazy('learner_dashboard:v0:program_progress_detail', kwargs={'program_uuid': program_uuid})
+    url = reverse_lazy("programs_api:v0:program_progress_detail", kwargs={"program_uuid": program_uuid})
 
     @classmethod
     def setUpClass(cls):
@@ -62,9 +63,9 @@ class TestProgramProgressDetailView(ProgramsApiConfigMixin, SharedModuleStoreTes
 
         cls.program_data = ProgramFactory(uuid=cls.program_uuid, courses=[course])
         cls.pathway_data = PathwayFactory()
-        cls.program_data['pathway_ids'] = [cls.pathway_data['id']]
-        cls.pathway_data['program_uuids'] = [cls.program_data['uuid']]
-        del cls.pathway_data['programs']  # lint-amnesty, pylint: disable=unsupported-delete-operation
+        cls.program_data["pathway_ids"] = [cls.pathway_data["id"]]
+        cls.pathway_data["program_uuids"] = [cls.program_data["uuid"]]
+        del cls.pathway_data["programs"]  # lint-amnesty, pylint: disable=unsupported-delete-operation
 
     def setUp(self):
         super().setUp()
@@ -74,25 +75,25 @@ class TestProgramProgressDetailView(ProgramsApiConfigMixin, SharedModuleStoreTes
 
     def assert_program_data_present(self, response):
         """Verify that program data is present."""
-        self.assertContains(response, 'program_data')
-        self.assertContains(response, 'course_data')
-        self.assertContains(response, 'urls')
-        self.assertContains(response, 'certificate_data')
-        self.assertContains(response, self.program_data['title'])
+        self.assertContains(response, "program_data")
+        self.assertContains(response, "course_data")
+        self.assertContains(response, "urls")
+        self.assertContains(response, "certificate_data")
+        self.assertContains(response, self.program_data["title"])
 
     def assert_pathway_data_present(self, response):
-        """ Verify that the correct pathway data is present. """
-        self.assertContains(response, 'industry_pathways')
-        self.assertContains(response, 'credit_pathways')
+        """Verify that the correct pathway data is present."""
+        self.assertContains(response, "industry_pathways")
+        self.assertContains(response, "credit_pathways")
 
-        industry_pathways = response.data['industry_pathways']
-        credit_pathways = response.data['credit_pathways']
-        if self.pathway_data['pathway_type'] == PathwayType.CREDIT.value:
-            credit_pathway, = credit_pathways  # Verify that there is only one credit pathway
+        industry_pathways = response.data["industry_pathways"]
+        credit_pathways = response.data["credit_pathways"]
+        if self.pathway_data["pathway_type"] == PathwayType.CREDIT.value:
+            (credit_pathway,) = credit_pathways  # Verify that there is only one credit pathway
             assert self.pathway_data == credit_pathway
             assert [] == industry_pathways
-        elif self.pathway_data['pathway_type'] == PathwayType.INDUSTRY.value:
-            industry_pathway, = industry_pathways  # Verify that there is only one industry pathway
+        elif self.pathway_data["pathway_type"] == PathwayType.INDUSTRY.value:
+            (industry_pathway,) = industry_pathways  # Verify that there is only one industry pathway
             assert self.pathway_data == industry_pathway
             assert [] == credit_pathways
 
@@ -104,11 +105,28 @@ class TestProgramProgressDetailView(ProgramsApiConfigMixin, SharedModuleStoreTes
         mock_get_programs.return_value = self.program_data
         mock_get_pathways.return_value = self.pathway_data
 
-        with mock.patch('lms.djangoapps.learner_dashboard.api.v0.views.get_certificates') as certs:
-            certs.return_value = [{'type': 'program', 'url': '/'}]
+        with mock.patch("openedx.core.djangoapps.programs.rest_api.v1.views.get_certificates") as certs:
+            certs.return_value = [{"type": "program", "url": "/"}]
             response = self.client.get(self.url)
 
-        assert response.status_code == 200
+        self.assertEqual(200, response.status_code)
+        self.assert_program_data_present(response)
+        self.assert_pathway_data_present(response)
+
+    def test_api_returns_correct_program_data(self, mock_get_programs, mock_get_pathways):
+        """
+        Verify that API returns program data in the correct format for URL format 2.
+        """
+        self.create_programs_config()
+        # fixme I think the problem is happening here? We are getting to get_programs and the return value is not including courses
+        mock_get_programs.return_value = self.program_data
+        mock_get_pathways.return_value = self.pathway_data
+
+        with mock.patch("openedx.core.djangoapps.programs.rest_api.v1.views.get_certificates") as certs:
+            certs.return_value = [{"type": "program", "url": "/"}]
+            response = self.client.get(self.url)
+
+        self.assertEqual(200, response.status_code)
         self.assert_program_data_present(response)
         self.assert_pathway_data_present(response)
 
@@ -135,7 +153,7 @@ class TestProgramProgressDetailView(ProgramsApiConfigMixin, SharedModuleStoreTes
 
         response = self.client.get(self.url)
         assert response.status_code == 404
-        assert response.data['error_code'] == 'No program data available.'
+        assert response.data["error_code"] == "No program data available."
 
 
 class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
@@ -143,7 +161,7 @@ class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
 
     enterprise_uuid = str(uuid4())
     program_uuid = str(uuid4())
-    url = reverse_lazy('learner_dashboard:v0:program_list', kwargs={'enterprise_uuid': enterprise_uuid})
+    url = reverse_lazy("programs_api:v0:program_list", kwargs={"enterprise_uuid": enterprise_uuid})
 
     @classmethod
     def setUpClass(cls):
@@ -155,30 +173,27 @@ class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
         course = CourseFactory(course_runs=[course_run])
         enterprise_customer = EnterpriseCustomerFactory(uuid=cls.enterprise_uuid)
         enterprise_customer_user = EnterpriseCustomerUserFactory(
-            user_id=cls.user.id,
-            enterprise_customer=enterprise_customer
+            user_id=cls.user.id, enterprise_customer=enterprise_customer
         )
-        CourseEnrollmentFactory(
-            is_active=True,
-            course_id=modulestore_course.id,
-            user=cls.user
-        )
+        CourseEnrollmentFactory(is_active=True, course_id=modulestore_course.id, user=cls.user)
         EnterpriseCourseEnrollmentFactory(
             course_id=modulestore_course.id,
-            enterprise_customer_user=enterprise_customer_user
+            enterprise_customer_user=enterprise_customer_user,
         )
 
         cls.program = ProgramFactory(
             uuid=cls.program_uuid,
             courses=[course],
-            title='Journey to cooking',
-            type='MicroMasters',
-            authoring_organizations=[{
-                'key': 'MAX',
-                'logo_image_url': 'http://test.org/media/organization/logos/test-logo.png'
-            }],
+            title="Journey to cooking",
+            type="MicroMasters",
+            authoring_organizations=[
+                {
+                    "key": "MAX",
+                    "logo_image_url": "http://test.org/media/organization/logos/test-logo.png",
+                }
+            ],
         )
-        cls.site = SiteFactory(domain='test.localhost')
+        cls.site = SiteFactory(domain="test.localhost")
 
     def setUp(self):
         super().setUp()
@@ -187,10 +202,10 @@ class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
         ProgramEnrollmentFactory.create(
             user=self.user,
             program_uuid=self.program_uuid,
-            external_user_key='0001',
+            external_user_key="0001",
         )
 
-    @with_site_configuration(configuration={'COURSE_CATALOG_API_URL': 'foo'})
+    @with_site_configuration(configuration={"COURSE_CATALOG_API_URL": "foo"})
     def test_program_list(self):
         """
         Verify API returns proper response.
@@ -198,7 +213,7 @@ class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
         cache.set(
             SITE_PROGRAM_UUIDS_CACHE_KEY_TPL.format(domain=self.site.domain),
             [self.program_uuid],
-            None
+            None,
         )
 
         response = self.client.get(self.url)
@@ -206,33 +221,31 @@ class TestProgramsView(SharedModuleStoreTestCase, ProgramCacheMixin):
         program = response.data[0]
 
         assert len(program)
-        assert program['uuid'] == self.program['uuid']
-        assert program['title'] == self.program['title']
-        assert program['type'] == self.program['type']
-        assert program['authoring_organizations'] == self.program['authoring_organizations']
-        assert program['banner_image'] == self.program['banner_image']
-        assert program['progress'] == {
-            'uuid': self.program['uuid'],
-            'completed': 0,
-            'in_progress': 0,
-            'not_started': 1,
-            'all_unenrolled': False,
+        assert program["uuid"] == self.program["uuid"]
+        assert program["title"] == self.program["title"]
+        assert program["type"] == self.program["type"]
+        assert program["authoring_organizations"] == self.program["authoring_organizations"]
+        assert program["banner_image"] == self.program["banner_image"]
+        assert program["progress"] == {
+            "uuid": self.program["uuid"],
+            "completed": 0,
+            "in_progress": 0,
+            "not_started": 1,
+            "all_unenrolled": False,
         }
 
-    @with_site_configuration(configuration={'COURSE_CATALOG_API_URL': 'foo'})
+    @with_site_configuration(configuration={"COURSE_CATALOG_API_URL": "foo"})
     def test_program_empty_list_if_no_enterprise_enrollments(self):
         """
         Verify API returns empty response if no enterprise enrollments exists for a learner.
         """
         # delete all enterprise course enrollments for the user
-        EnterpriseCourseEnrollment.objects.filter(
-            enterprise_customer_user__user_id=self.user.id
-        ).delete()
+        EnterpriseCourseEnrollment.objects.filter(enterprise_customer_user__user_id=self.user.id).delete()
 
         cache.set(
             SITE_PROGRAM_UUIDS_CACHE_KEY_TPL.format(domain=self.site.domain),
             [self.program_uuid],
-            None
+            None,
         )
 
         response = self.client.get(self.url)
